control_plane: fix consistency check
